split block.py into block.py and block_manager.py
