+++ asdf/asdf.py
@@ -13,7 +13,7 @@ from packaging.version import Version
 from . import _display as display
 from . import _node_info as node_info
 from . import _version as version
-from . import block, constants, generic_io, reference, schema, treeutil, util, versioning, yamlutil
+from . import block_manager, constants, generic_io, reference, schema, treeutil, util, versioning, yamlutil
 from ._helpers import validate_version
 from .config import config_context, get_config
 from .exceptions import AsdfConversionWarning, AsdfDeprecationWarning, AsdfWarning
@@ -156,7 +156,7 @@ class AsdfFile:
         self._fd = None
         self._closed = False
         self._external_asdf_by_uri = {}
-        self._blocks = block.BlockManager(self, copy_arrays=copy_arrays, lazy_load=lazy_load, readonly=_readonly)
+        self._blocks = block_manager.BlockManager(self, copy_arrays=copy_arrays, lazy_load=lazy_load, readonly=_readonly)
         self._uri = None
         if tree is None:
             # Bypassing the tree property here, to avoid validating
@@ -1242,7 +1242,7 @@ class AsdfFile:
 
                 serialized_tree_size = tree_serialized.tell() + constants.MAX_BLOCKS_DIGITS * array_ref_count[0]
 
-                if not block.calculate_updated_layout(self.blocks, serialized_tree_size, pad_blocks, fd.block_size):
+                if not block_manager.calculate_updated_layout(self.blocks, serialized_tree_size, pad_blocks, fd.block_size):
                     # If we don't have any blocks that are being reused, just
                     # write out in a serial fashion.
                     self._serial_write(fd, pad_blocks, include_block_index)
